all the files from prepro are moved
