+++ raw-data/c++/cc.cpp
