+++ raw-data/c++/dd.cpp
